refactor(assignment): migrate module
