+++ lib/app/app.dart
@@ -8,7 +8,7 @@ export 'package:hive/hive.dart'
 export 'package:hive_cache/hive_cache.dart';
 export 'package:meta/meta.dart';
 export 'package:pedantic/pedantic.dart';
-export 'package:time_machine/time_machine.dart' show Instant;
+export 'package:time_machine/time_machine.dart' show Instant, LocalDate, Period;
 
 export 'app_config.dart';
 export 'data.dart';
