+++ lib/assignment/assignment.dart
@@ -1,4 +1,3 @@
 export 'data.dart';
 export 'routes.dart';
-export 'widgets/assignments_screen.dart';
 export 'widgets/dashboard_card.dart';
