+++ lib/assignment/widgets/grade_indicator.dart
@@ -1,5 +1,5 @@
-import 'package:black_hole_flutter/black_hole_flutter.dart';
 import 'package:flutter/material.dart';
+import 'package:schulcloud/app/app.dart';
 
 import '../data.dart';
