+++ lib/l10n/intl_de.arb
@@ -55,13 +55,12 @@
   "assignment_assignmentDetails_submissions_placeholder": "In diesem Tab siehst du bald die Abgaben von allen Schülern.",
   "assignment_assignmentDetails_unarchive": "Wiederherstellen",
   "assignment_assignmentDetails_unarchived": "Wiederhergestellt",
-  "assignment_assignmentsScreen_empty": "Du hast keine Aufgaben.",
-  "assignment_assignmentsScreen_emptyFiltered": "Es wurden keine Aufgaben gefunden, die deinen Filtern entsprechen.",
+  "assignment_assignmentsPage_empty": "Du hast keine Aufgaben.",
+  "assignment_assignmentsPage_emptyFiltered": "Es wurden keine Aufgaben gefunden, die deinen Filtern entsprechen.",
   "assignment_dashboardCard": "Hausaufgaben",
   "assignment_dashboardCard_all": "Alle Hausaufgaben",
   "assignment_dashboardCard_header": "{assignmentCount, plural, one{offene Hausaufgabe\nin der nächsten Woche} other{offene Hausaufgaben\nin der nächsten Woche}}",
   "assignment_dashboardCard_noCourse": "(ohne Kurs)",
-  "assignment_detailsScreen_mySubmission": "Meine Abgabe",
   "assignment_editSubmission_delete_confirm": "Möchtest du diese Abgabe wirklich löschen?",
   "assignment_editSubmission_delete_success": "Erfolgreich gelöscht",
   "assignment_editSubmission_overwriteFormatting": "Durch das Bearbeiten dieser Abgabe geht die vorhandene Formatierung verloren.",
