+++ lib/l10n/intl_en.arb
@@ -55,13 +55,12 @@
   "assignment_assignmentDetails_submissions_placeholder": "You will soon be able to see submissions of all students in this tab.",
   "assignment_assignmentDetails_unarchive": "Unarchive",
   "assignment_assignmentDetails_unarchived": "Unarchived",
-  "assignment_assignmentsScreen_empty": "You don't have any assignments.",
-  "assignment_assignmentsScreen_emptyFiltered": "No assignments found matching your filter criteria.",
+  "assignment_assignmentsPage_empty": "You don't have any assignments.",
+  "assignment_assignmentsPage_emptyFiltered": "No assignments found matching your filter criteria.",
   "assignment_dashboardCard": "Assignments",
   "assignment_dashboardCard_all": "All assignments",
   "assignment_dashboardCard_header": "{assignmentCount, plural, one{Open assignment\nin the next week} other{Open assignments\nin the next week}}",
   "assignment_dashboardCard_noCourse": "(without course)",
-  "assignment_detailsScreen_mySubmission": "My submission",
   "assignment_editSubmission_delete_confirm": "Dou you really want to delete this submission?",
   "assignment_editSubmission_delete_success": "Deleted successfully",
   "assignment_editSubmission_overwriteFormatting": "Editing this submission will remove all existing formatting.",
